refactor(sprint14): external_imap/mod.rs (1303→400) split par domaine
- account_ops.rs (134) : create/list/get/update/delete_account + status
- folder_ops.rs (98) : list_folders + upsert_folder_mapping + ensure_folder
- sync_ops.rs (102) : start/complete/get_sync_run/get_sync_status
- message_ops.rs (113) : list_messages + apply_message_action
- imap_client_ops.rs (493) : imap_test + discover_folders + run_sync_now + helpers réseau
- mod.rs : structs publics + setup (new/db_name/coll_*)
